% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAM-package.R
\docType{package}
\name{DTAM-package}
\alias{DTAM}
\alias{DTAM-package}
\title{DTAM: Discrete-Time Affine Models}
\description{
Simulation, filtering, estimation, and pricing tools for discrete-time
affine asset-pricing models.
}
\author{
\strong{Maintainer}: Jean-Paul Renne \email{jean-paul.renne@unil.ch}

}
\keyword{internal}
