% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-aliases.R
\name{legacy_aliases}
\alias{legacy_aliases}
\alias{reverse.MHLT}
\alias{psi.GaussianVAR}
\alias{psi.GaussianQVAR}
\alias{psi.VARG}
\alias{psi.VARG_Poisson}
\alias{psi.QPoisson}
\alias{psi.RSVAR}
\alias{psi.BansalShaliastovich}
\alias{psi.AR_CH}
\alias{psi.w.TopDown}
\alias{psiQ.w.TopDown}
\alias{psi.y.TopDown}
\alias{simul.ARG}
\alias{simul.GVAR}
\alias{simul.RSVAR}
\alias{simul.TopDown}
\alias{simul.compound.poisson}
\alias{simul.rating.migration}
\alias{make.pdf}
\title{Legacy dot-style aliases}
\usage{
reverse.MHLT(...)

psi.GaussianVAR(...)

psi.GaussianQVAR(...)

psi.VARG(...)

psi.VARG_Poisson(...)

psi.QPoisson(...)

psi.RSVAR(...)

psi.BansalShaliastovich(...)

psi.AR_CH(...)

psi.w.TopDown(...)

psiQ.w.TopDown(...)

psi.y.TopDown(...)

simul.ARG(...)

simul.GVAR(...)

simul.RSVAR(...)

simul.TopDown(...)

simul.compound.poisson(...)

simul.rating.migration(...)

make.pdf(...)
}
\arguments{
\item{...}{Arguments passed to the underscore-named replacement.}
}
\description{
These wrappers preserve backward compatibility with earlier DTAM code that
used dot-style function names. New code should use the underscore-named
replacements documented throughout the package.
}
