% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedures.R
\name{make.pdf}
\alias{make.pdf}
\title{Recover a density from a Laplace transform}
\usage{
make.pdf(
  model,
  values.of.variable,
  psi,
  max_x = 10000,
  min_x = 10^(-8),
  nb_x = 5000
)
}
\arguments{
\item{model}{List of model parameters passed to \code{psi}.}

\item{values.of.variable}{Grid at which the density is approximated.}

\item{psi}{Function returning the Laplace transform at the requested points.}

\item{max_x, min_x, nb_x}{Numerical integration settings for the inversion.}
}
\value{
Numeric vector of approximate density values on
\code{values.of.variable}.
}
\description{
Uses Fourier inversion to approximate a univariate density on a user-supplied
grid of values.
}
\examples{
psi_exp <- function(u, model) {
  -log(1 - model$mu * u)
}
grid <- seq(0.01, 2, length.out = 50)
dens <- make.pdf(list(mu = 1), values.of.variable = grid, psi = psi_exp,
                 max_x = 100, nb_x = 500)
length(dens)

}
