% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive-preferences.R
\name{psi.BansalShaliastovich}
\alias{psi.BansalShaliastovich}
\title{Log-Laplace transform for the Bansal-Shaliastovich model}
\usage{
psi.BansalShaliastovich(u, psi.parameterization)
}
\arguments{
\item{u}{Matrix of transform arguments.}

\item{psi.parameterization}{List containing the model matrices and loadings.}
}
\value{
A list with affine-transform coefficients \code{a} and \code{b}.
}
\description{
Evaluates the conditional log-Laplace transform used in the long-run-risk
specification of Bansal and Shaliastovich.
}
\examples{
model <- list(
  Pi = diag(c(0.9, 0.8)),
  nu = matrix(c(0.7), 1, 1),
  chi.0 = c(0.2, 0.3),
  chi.1 = matrix(c(0.5, 0.2), 1, 2),
  sigma.w = 0.2
)
u <- matrix(c(0.1, 0.05, 0.02), 3, 1)
res <- psi.BansalShaliastovich(u, model)
dim(res$a)

}
