% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow-rate.R
\name{psi_non_gaussian}
\alias{psi_non_gaussian}
\alias{psi.QPoisson}
\alias{psi.VARG_Poisson}
\alias{psi.AR_CH}
\title{Laplace transforms for non-Gaussian auxiliary models}
\usage{
psi.QPoisson(u, psi.parameterization)

psi.VARG_Poisson(u, psi.parameterization)

psi.AR_CH(u, psi.parameterization)
}
\arguments{
\item{u}{Matrix of transform arguments, one column per evaluation point.}

\item{psi.parameterization}{List of model parameters.}
}
\value{
A list with affine-transform coefficients \code{a} and \code{b}.
}
\description{
These functions return affine or log-Laplace-transform coefficients for a
set of non-Gaussian state dynamics used in the book: quadratic-Gaussian
Poisson counts, VARG-Poisson models, and autoregressive gamma-volatility
models.
}
\examples{
model_ar_ch <- list(mu = 0, phi = 0.9, nu = 1, mu_z = 0.2, beta = 0.8)
res <- psi.AR_CH(matrix(c(0.1, 0.05), 2, 1), model_ar_ch)
dim(res$a)

}
