% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAM-package.R
\docType{package}
\name{DTAM-package}
\alias{DTAM}
\alias{DTAM-package}
\title{DTAM: Discrete-Time Affine Models}
\description{
Simulation, filtering, estimation, and pricing tools for discrete-time
affine asset-pricing models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jrenne/DTAM}
  \item \url{https://jrenne.github.io/MPRR_codes/}
  \item Report bugs at \url{https://github.com/jrenne/DTAM/issues}
}

}
\author{
\strong{Maintainer}: Jean-Paul Renne \email{jean-paul.renne@unil.ch}

}
\keyword{internal}
