% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{make_recessions}
\alias{make_recessions}
\title{Add recession shading to a time-series plot}
\usage{
make_recessions(col = "#AA55AA44", indic_US = TRUE, MIN = -10000, MAX = +10000)
}
\arguments{
\item{col}{Fill color used for the shaded areas.}

\item{indic_US}{Logical. If \code{TRUE}, use U.S. recession dates from \code{tseries};
otherwise use the hard-coded euro-area dates included in the package.}

\item{MIN, MAX}{Vertical bounds used for the polygon.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Draws shaded polygons corresponding to U.S. or euro-area recession dates on
the current base-R plot.
}
\examples{
plot(as.Date(c("2007-01-01", "2012-01-01")),
     c(0, 1), type = "n", xlab = "", ylab = "")
make_recessions(col = "#CCCCCC66", MIN = 0, MAX = 1)

}
